Create setup function
