+++ signet-program/tests/chain-signatures-project.ts
@@ -1,65 +1,16 @@
 import * as anchor from "@coral-xyz/anchor";
-import { BN, Program } from "@coral-xyz/anchor";
-import type { ChainSignaturesProject } from "../target/types/chain_signatures_project";
-import { MockSignerServer } from "../test-utils/MockSignerServer";
-import { Connection } from "@solana/web3.js";
+import { BN } from "@coral-xyz/anchor";
 import { assert } from "chai";
-import { SignatureRespondedSubscriber } from "../test-utils/SignatureRespondedSubscriber";
-import { bigintPrivateKeyToNajPublicKey, getEnv } from "../test-utils/utils";
-import { chainAdapters, contracts } from "signet.js";
+import { setup } from "../test-utils/setup";
 
 describe("chain-signatures-project", () => {
-  const provider = anchor.AnchorProvider.env();
-  anchor.setProvider(provider);
-
-  const connection = new Connection(provider.connection.rpcEndpoint);
-
-  const program = anchor.workspace
-    .chainSignaturesProject as Program<ChainSignaturesProject>;
-
-  const env = getEnv();
-
-  const rootPublicKey = bigintPrivateKeyToNajPublicKey(env.PRIVATE_KEY_TESTNET);
-
-  const signetSolContract = new contracts.solana.ChainSignatureContract({
+  const {
     provider,
-    programId: program.programId,
-    rootPublicKey,
-  });
-
-  const mockServer = new MockSignerServer({ provider, signetSolContract });
-
-  const evmChainAdapter = new chainAdapters.evm.EVM({
-    publicClient: {} as any, // Don't care, EVM chain adapter only used to derive address
-    contract: signetSolContract,
-  });
-
-  const signatureRespondedSubscriber = new SignatureRespondedSubscriber(
-    program
-  );
-
-  before(async () => {
-    const tx = await program.methods.initialize(new BN("100000")).rpc();
-
-    const latestBlockhash = await connection.getLatestBlockhash();
-
-    await connection.confirmTransaction(
-      {
-        signature: tx,
-        blockhash: latestBlockhash.blockhash,
-        lastValidBlockHeight: latestBlockhash.lastValidBlockHeight,
-      },
-      "confirmed"
-    );
-  });
-
-  beforeEach(async () => {
-    await mockServer.start();
-  });
-
-  afterEach(async () => {
-    await mockServer.stop();
-  });
+    program,
+    signetSolContract,
+    evmChainAdapter,
+    signatureRespondedSubscriber,
+  } = setup();
 
   it("Is initialized!", async () => {
     const [programStatePda] = anchor.web3.PublicKey.findProgramAddressSync(
